Translate unit tests to TypeScript
